doc/ref/spec.md: document new ... operator
This documents the new closedness behavior using
the ... operator.

Note that this replaces the old documentation. As of
now, this is just an experiment. We might want to
document this separately on a separate page.

It also defines the `__closeAll` builtin, which is
used by `cue fix` to convert legacy code to the
new semantics.

Discussion #4032

Signed-off-by: Marcel van Lohuizen <mpvl@gmail.com>
Change-Id: I6206917ffee4f6f936b03199231a004922f3be15
GitOrigin-RevId: d763a081f95e762a3806a38924de2a145a77a61f